pyproject.toml
[tool.pyright]
venvPath = "."
venv = ".venv"

[project]
name = "backend"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.18",
    "argon2-cffi>=23.1.0",
    "fastapi>=0.115.12",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "python-decouple>=3.8",
    "supabase>=2.15.1",
    "uvicorn[standard]>=0.34.2",
]
